+++ integration_tests/test_drawing_matplotlib_backend.py
@@ -1,0 +1,18 @@
+# Created: 06.2020
+# Copyright (c) 2020, Matthew Broadway
+# License: MIT License
+import pytest
+
+plt = pytest.importorskip('matplotlib.pyplot')
+
+from ezdxf.addons.drawing.matplotlib_backend import MatplotlibBackend
+
+
+@pytest.fixture()
+def backend():
+    fig, ax = plt.subplots()
+    return MatplotlibBackend(ax)
+
+
+def test_get_text_width(backend):
+    assert backend.get_text_line_width('   abc', 100) > backend.get_text_line_width('abc', 100)